renaming keyhashes to commitments
