+++ prover/vk_setup_data_generator_server_fri/historical_data/18/commitments.json
@@ -1,0 +1,6 @@
+{
+    "leaf": "0x72167c43a46cf38875b267d67716edc4563861364a3c03ab7aee73498421e828",
+    "node": "0x5a3ef282b21e12fe1f4438e5bb158fc5060b160559c5158c6389d62d9fe3d080",
+    "scheduler": "0x201d4c7d8e781d51a3bbd451a43a8f45240bb765b565ae6ce69192d918c3563d",
+    "snark_wrapper": "0x4be443afd605a782b6e56d199df2460a025c81b3dea144e135bece83612563f2"
+}
