--- prover/vk_setup_data_generator_server_fri/historical_data/18/key_hashes.json
@@ -1,3 +1,0 @@
-{
-    "snark_wrapper_vk_hash": "0x4be443afd605a782b6e56d199df2460a025c81b3dea144e135bece83612563f2"
-}
