+++ prover/vk_setup_data_generator_server_fri/historical_data/19/commitments.json
@@ -1,0 +1,6 @@
+{
+    "leaf": "0x14628525c227822148e718ca1138acfc6d25e759e19452455d89f7f610c3dcb8",
+    "node": "0x5a3ef282b21e12fe1f4438e5bb158fc5060b160559c5158c6389d62d9fe3d080",
+    "scheduler": "0x61ed1ea338680a9f83945b300a4a4778dd76d74a015b43a3e9b7fedba914a3d0",
+    "snark_wrapper": "0x750d8e21be7555a6841472a5cacd24c75a7ceb34261aea61e72bb7423a7d30fc"
+}
