--- prover/vk_setup_data_generator_server_fri/historical_data/19/key_hashes.json
@@ -1,3 +1,0 @@
-{
-    "snark_wrapper_vk_hash": "0x750d8e21be7555a6841472a5cacd24c75a7ceb34261aea61e72bb7423a7d30fc"
-}
