+++ prover/vk_setup_data_generator_server_fri/historical_data/20/commitments.json
@@ -1,0 +1,6 @@
+{
+    "leaf": "0x062362cb3eaf1f631406cbe19bf2a2c5d0d9ea69d069309a6003addae9f387be",
+    "node": "0x5a3ef282b21e12fe1f4438e5bb158fc5060b160559c5158c6389d62d9fe3d080",
+    "scheduler": "0xf12256f7ba1ab223ba82abb19c8366c6677596897f9e5a82016977113e52ed5d",
+    "snark_wrapper": "0x8574e152c41dc39a2ecab984545e1cf21cb3ec250b919018a8053f2fa270784f"
+}
