--- prover/vk_setup_data_generator_server_fri/historical_data/20/key_hashes.json
@@ -1,3 +1,0 @@
-{
-    "snark_wrapper_vk_hash": "0x8574e152c41dc39a2ecab984545e1cf21cb3ec250b919018a8053f2fa270784f"
-}
