+++ prover/vk_setup_data_generator_server_fri/src/main.rs
@@ -23,8 +23,10 @@ use zksync_prover_fri_types::{
     ProverServiceDataKey,
 };
 use zksync_vk_setup_data_server_fri::{
+    commitment_utils::generate_commitments,
     keystore::Keystore,
     setup_data_generator::{CPUSetupDataGenerator, GPUSetupDataGenerator, SetupDataGenerator},
+    utils::calculate_snark_vk_hash,
 };
 
 mod commitment_generator;
@@ -42,12 +44,15 @@ fn generate_vks(keystore: &Keystore) -> anyhow::Result<()> {
     tracing::info!("Generating verification keys for Base layer.");
     generate_base_layer_vks_and_proofs(&mut in_memory_source)
         .map_err(|err| anyhow::anyhow!("Failed generating base vk's: {err}"))?;
+
+    // This must happen before we start generating recursive layer.
+    tracing::info!("Generating 4844 base layer.");
+    generate_eip4844_vks(&mut in_memory_source)
+        .map_err(|err| anyhow::anyhow!("Failed generating 4844 vk's: {err}"))?;
+
     tracing::info!("Generating verification keys for Recursive layer.");
     generate_recursive_layer_vks_and_proofs(&mut in_memory_source)
         .map_err(|err| anyhow::anyhow!("Failed generating recursive vk's: {err}"))?;
-
-    generate_eip4844_vks(&mut in_memory_source)
-        .map_err(|err| anyhow::anyhow!("Failed generating 4844 vk's: {err}"))?;
 
     tracing::info!("Saving keys & hints");
 
@@ -65,7 +70,10 @@ fn generate_vks(keystore: &Keystore) -> anyhow::Result<()> {
     let (_, vk) = get_wrapper_setup_and_vk_from_scheduler_vk(scheduler_vk, config);
     keystore
         .save_snark_verification_key(vk)
-        .context("save_snark_vk")
+        .context("save_snark_vk")?;
+
+    // Let's also update the commitments file.
+    keystore.save_commitments(&generate_commitments(keystore)?)
 }
 
 #[derive(Debug, Parser)]
